+++ CHANGELOG.md
@@ -4,6 +4,11 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [Unreleased]
+#### Changed
+- `@rarimo/rarime-connector` - bump version to 1.0.0-rc.0
+- Getting identity id's
+
 ## [2.3.0] - 2023-11-17
 #### Changed
 - env files interaction, now .env files are able to bundle in ci